pyproject.toml
[project]
name = "formation-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.128.0",
    "ffmpeg-python>=0.2.0",
    "hypercorn>=0.18.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "alembic>=1.18.1",
]
